wallet_service/pyproject.toml
[project]
name = "wallet-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic==1.17.2",
    "asyncpg==0.31.0",
    "fastapi==0.121.2",
    "psycopg2-binary==2.9.11",
    "pydantic==2.12.5",
    "pydantic-settings==2.12.0",
    "python-jose[cryptography]==3.5.0",
    "sqlalchemy==2.0.44",
    "uvicorn==0.38.0",
]
